♻️ Move tests to src/
Keeping tests next to the source files helps to find the corresponding
tests quicker, spot missing tests, and it results in simpler imports.
