+++ src/test/resources/liberty.jpg
