+++ src/test/resources/torus.png
